Proper build error handling
Exception raised from BuildContext.close were hiding original
exception, if any. Moved BuildContext.close outside of finally
block.

Wrap BuildException in MojoExecutionException, which has special
handling.

As a side effect of this change, build state is not persisted in
case of exception in mojo. Build state is likely in inconsistent
unknown state this case, so not persisting such state is a good
thing.

Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>